CloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
pandas 
requests
datetime
pandas-gbq
google-auth